Move selector to the conflicts file
